requirements.txt
SQLAlchemy==2.0.9
python-telegram-bot==13.15
environs==9.5.0
qrcode==7.4.2
requests==2.28.2